+++ src/baobab_scryfall_api_caller/client/__init__.py
@@ -2,5 +2,6 @@
 
 from baobab_scryfall_api_caller.client.scryfall_api_caller import ScryfallApiCaller
 from baobab_scryfall_api_caller.client.scryfall_http_client import ScryfallHttpClient
+from baobab_scryfall_api_caller.client.web_api_transport_protocol import WebApiTransportProtocol
 
-__all__ = ["ScryfallApiCaller", "ScryfallHttpClient"]
+__all__ = ["ScryfallApiCaller", "ScryfallHttpClient", "WebApiTransportProtocol"]